+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
 ## 0.3.4
 - Allow adding folders from the DM Binder view
+- Allow adding files from the DM Binder view
 
 ## 0.3.3
 - Make sure the dmbinder.autogenerateOnRender setting actually does something (whoops)